rename components package to layout
